Extract SimpleHookConfig
