+++ app/src/main/java/andhook/test/ui/MainActivity.java
@@ -15,9 +15,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import andhook.lib.AndHook;
-import andhook.lib.HookHelper;
-import andhook.lib.HookHelper.Hook;
 import andhook.test.AndTest;
+import andhook.test.SimpleHookConfig;
 import andhook.test.Constructor;
 import andhook.test.GC;
 import andhook.test.InnerException;
@@ -35,25 +34,10 @@ import andhook.test.app.MainApplication;
 @SuppressWarnings("all")
 public class MainActivity extends Activity {
     private static final String TAG = AndTest.LOG_TAG;
-    private static boolean passed = false;
     private static MainActivity thiz = null;
     private static EditText tv_status = null;
     private static TextView tv_more = null;
     private static CharSequence cv_more = null;
-
-    /**
-     * 对 Activity 的 onCreate(Bunlde); 方法进行 Hook
-     *
-     * @param objActivity        被 Hook 的 activity 对象
-     * @param savedInstanceState onCreate(Bunlde) 的 bundle 入参
-     */
-    @Hook(clazz = Activity.class, name = "onCreate")
-    private static void Activity_onCreate(final Object objActivity, final Bundle savedInstanceState) {
-        Log.i(AndTest.LOG_TAG, "HookedActivity::onCreate start, this is " + objActivity.getClass());
-        HookHelper.invokeVoidOrigin(objActivity, savedInstanceState);
-        Log.i(AndTest.LOG_TAG, "HookedActivity::onCreate end");
-        passed = true;
-    }
 
     public static void runAction(final Runnable action) {
         if (thiz != null)
@@ -104,13 +88,6 @@ public class MainActivity extends Activity {
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
-        // 从指定的配置类中加载被 @Hook 标记的方法.
-        // 本例演示的是加载 MainActivity 中的 Activity_onCreate(Activity, Bundle);
-        Log.d(TAG, "MainActivity::onCreate: before apply @Hook");
-        HookHelper.applyHooks(MainActivity.class);
-        Log.d(TAG, "MainActivity::onCreate: after apply @Hook");
-
-
         Log.i(AndTest.LOG_TAG, "MainActivity.super::onCreate start");
         super.onCreate(savedInstanceState);
         Log.i(AndTest.LOG_TAG, "MainActivity.super::onCreate end");
@@ -143,7 +120,7 @@ public class MainActivity extends Activity {
         clear();
         output(AndHook.class + " version " + AndHook.VERSION + " ("
                 + AndHook.getVersionInfo() + ")");
-        if (!passed)
+        if (!SimpleHookConfig.passed)
             alert("Activity::onCreate hook failed!");
 
         findViewById(R.id.JNI).setOnClickListener(new OnClickListener() {
